# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ParamCommunity:
    properties:
      introduction:
        type: string
      name:
        type: string
    type: object
...
  models.ParamLoginUsingEmail:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.ParamLoginUsingPhoneWithCode:
    properties:
      code:
        type: string
      phone:
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: 1298453249@qq.com
    name: xiaorui zheng
    url: http://www.swagger.io/support
  description: 这是一个热点论坛项目， 能够根据当下热点来向用户展示论坛帖子
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 热点论坛
  version: "1.0"
paths:
  /auth//login/refresh-token:
    get:
      consumes:
      - application/json
      description: 使用邮箱+密码的方式进行登陆
      parameters:
      - description: 刷新Token
        in: query
        name: refresh_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 使用邮箱+密码的方式进行登陆
      tags:
      - Auth
...
  /auth/login/email:
    post:
      consumes:
      - application/json
      description: 使用邮箱+密码的方式进行登陆
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamLoginUsingEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 使用邮箱+密码的方式进行登陆
      tags:
      - Auth
  /auth/login/phone:
    post:
      consumes:
      - application/json
      description: 实现使用手机号码进行登陆的功能
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamLoginUsingPhoneWithCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 实现使用手机号码进行登陆的功能
      tags:
      - Auth
...
  /community:
    get:
      consumes:
      - application/json
      description: 创建新的社区
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建新的社区
      tags:
      - Community
    post:
      consumes:
      - application/json
      description: 创建新的社区
      parameters:
      - description: 参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamCommunity'
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建新的社区
      tags:
      - Community
  /community/{id}:
    delete:
      consumes:
      - application/json
      description: 删除某个社区的信息
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除某个社区的信息
      tags:
      - Community
    get:
      consumes:
      - application/json
      description: 创建新的社区
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建新的社区
      tags:
      - Community
    put:
      consumes:
      - application/json
      description: 更新某个社区的信息
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新某个社区的信息
      tags:
      - Community
...
  /post/comment/{comment_id}:
    delete:
      consumes:
      - application/json
      description: 删除某条评论
      parameters:
      - description: comment ID
        in: path
        name: comment_id
        required: true
        type: integer
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除某条评论
      tags:
      - Comment
  /post/comment/{post_id}:
    get:
      consumes:
      - application/json
      description: 获得某个post下的所有评论
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: Page number
        in: query
        name: page_num
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获得某个post下的所有评论
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: 创建一个评论
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建一个评论
      tags:
      - Comment
...
swagger: "2.0"